Moving all configuration related file into 1 package, and extracting all like minded code from GridHubConfiguration into a single instance of ConfigurationLoader.java
