+++ java/server/src/org/openqa/grid/web/servlet/RegistrationServlet.java
@@ -30,7 +30,7 @@ import org.openqa.grid.common.RegistrationRequest;
 import org.openqa.grid.internal.BaseRemoteProxy;
 import org.openqa.grid.internal.Registry;
 import org.openqa.grid.internal.RemoteProxy;
-import org.openqa.grid.internal.utils.GridHubConfiguration;
+import org.openqa.grid.internal.utils.configuration.GridHubConfiguration;
 
 /**
  * entry point for the registration API the grid provides. The {@link RegistrationRequest} sent to
