+++ java/server/test/org/openqa/grid/internal/UserDefinedCapabilityMatcherTests.java
@@ -22,7 +22,7 @@ import org.junit.Assert;
 import org.junit.Test;
 import org.openqa.grid.common.RegistrationRequest;
 import org.openqa.grid.internal.utils.DefaultCapabilityMatcher;
-import org.openqa.grid.internal.utils.GridHubConfiguration;
+import org.openqa.grid.internal.utils.configuration.GridHubConfiguration;
 import org.openqa.grid.web.Hub;
 
 public class UserDefinedCapabilityMatcherTests {
